+++ src/components/course/list/lecturer/LecturerCourse.vue
@@ -27,7 +27,7 @@
                         <read-more more-str="Show more" :text="course.courseDescription" less-str="Show less" :max-chars="180"></read-more>
                     </div>
                 </div>
-                <div v-if="course.lecturerId == username" class="w-1/3 ml-12 lg:w-1/6 lg:ml-8">
+                <div v-if="allowEdit" class="w-1/3 ml-12 lg:w-1/6 lg:ml-8">
                     <div class="flex items-center justify-center h-full mt-6">
                         <button id="editCourse" class="w-48 py-2 btn btn-gray-primary" @click="editCourse()">Edit</button>
                     </div>
@@ -59,8 +59,8 @@
                 type: Object as () => Lecturer,
                 required: true,
             },
-            username: {
-                type: String,
+            allowEdit: {
+                type: Boolean,
                 required: true,
             },
         },