+++ src/use/router/index.ts
@@ -1,9 +1,9 @@
 import { createRouter, createWebHistory } from "vue-router";
 import LoginView from "@/views/common/Login.vue";
 import StudentCourseView from "@/views/student/StudentCourseList.vue";
-import LecturerCourseView from "@/views/lecturer/LecturerCourseList.vue";
+import AllCourseView from "@/views/shared/CourseList.vue";
 import AdminAccountListView from "@/views/admin/AdminAccountList.vue";
-import CourseFormSuspenseWrapper from "@/views/lecturer/EditCreateCourseForm.vue";
+import CourseFormSuspenseWrapper from "@/views/shared/EditCreateCourseForm.vue";
 import AccountFormSuspenseWrapper from "@/views/admin/EditCreateAccountForm.vue";
 import Redirect from "@/views/errors/403.vue";
 import ProfileWrapper from "@/components/profile/Wrapper.vue";
@@ -60,7 +60,7 @@ const router = createRouter({
             path: "/course-management",
             name: "lecturer.myCourses",
             props: { showAllCourses: false },
-            component: LecturerCourseView,
+            component: AllCourseView,
             meta: {
                 title: "My Courses" + suffix,
             },
@@ -68,12 +68,12 @@ const router = createRouter({
 
         {
             path: "/all-courses",
-            name: "lecturer.courses",
+            name: "shared.courses",
             props: { showAllCourses: true },
-            component: LecturerCourseView,
+            component: AllCourseView,
             meta: {
                 title: "All Courses" + suffix,
-                roles: ["Lecturer"],
+                roles: ["Lecturer", "Admin"],
             },
         },
 
@@ -96,7 +96,7 @@ const router = createRouter({
             component: CourseFormSuspenseWrapper,
             meta: {
                 title: "Course Creation" + suffix,
-                roles: ["Lecturer"],
+                roles: ["Lecturer", "Admin"],
             },
         },
         {
@@ -108,7 +108,7 @@ const router = createRouter({
             component: CourseFormSuspenseWrapper,
             meta: {
                 title: "Course Editing" + suffix,
-                roles: ["Lecturer"],
+                roles: ["Lecturer", "Admin"],
             },
         },
         {